Clearer division between common and helpers
Moving some of the functions from one to the other. Leaving
it backwards compatible though (with @deprecated).
